requirements.txt
Django==2.2.6
gunicorn==20.0.2
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
python-dateutil==2.8.1
